Added systemd unit for setting charge threshold on startup
I have my laptop plugged in most of the time, so I don't want it to sit
a 100% battery all the time.

I chose 80 as the upper and 40 as the lower threshold - just because
that is what I used to have on my ThinkPad. Not sure if it is the best
setting.