Workaround shortcut; Window -> Monitor assigment not working
When having my monitor connected, the workspace asignemnt does not work
for it. It always starts with the 11th workspace being on the monitor.
So I need this wacky shortcut for now, so that I can access this
workspace until I debugged why my monitor does not start with the 101th
workspace